feat(agent): 支持 GLM-5.2 / GLM-X-Preview[1m] 的 1M 上下文窗口
- zhipu / zhipu-coding 渠道默认模型列表追加 GLM-5.2（默认启用）和 GLM-X-Preview[1m]，5.1 保留
- zhipu-coding 连通性测试改用 glm-5.2
- 把 supports1MContext 和 inferContextWindow 抽到 packages/shared 作为单一 source of truth，顺便修掉前端 Sonnet 4/4.5、MiniMax-M3 与后端列表不一致的旧问题
