+++ packages/shared/src/utils/index.ts
@@ -10,6 +10,12 @@ export function noop(): void {
 export { diffCapabilities } from './capabilities-diff'
 export type { CapabilityChange } from './capabilities-diff'
 export {
+  DEFAULT_CONTEXT_WINDOW,
+  ONE_MILLION_CONTEXT_WINDOW,
+  supports1MContext,
+  inferContextWindow,
+} from './context-window'
+export {
   THINKING_SIGNATURE_ERROR_CODE,
   THINKING_SIGNATURE_ERROR_TITLE,
   THINKING_SIGNATURE_ERROR_MESSAGE,
